adding tests and create a facebook file test
